Add query option hooks for re-usability
