+++ domains/blocks/src/services/apollo/queries/attendees/index.ts
@@ -1,0 +1,1 @@
+export { default as useAttendees } from './useAttendees';
