+++ domains/blocks/src/services/apollo/queries/datetimes/index.ts
@@ -1,0 +1,1 @@
+export { default as useDatetimes } from './useDatetimes';
