+++ domains/blocks/src/services/apollo/queries/events/index.ts
@@ -1,0 +1,1 @@
+export { default as useEvents } from './useEvents';
