+++ domains/blocks/src/services/apollo/queries/events/useEvents.ts
@@ -1,0 +1,12 @@
+import type { FetchQueryResult } from '@eventespresso/data';
+import { useEventsQuery } from '@eventespresso/data';
+
+import { EventsList } from '@blocksServices/apollo/types';
+import useEventsQueryOptions from './useEventsQueryOptions';
+
+const useEvents = (): FetchQueryResult<EventsList> => {
+	const queryOptions = useEventsQueryOptions();
+	return useEventsQuery(queryOptions);
+};
+
+export default useEvents;
