+++ domains/blocks/src/services/apollo/queries/index.ts
@@ -1,4 +1,4 @@
-export { default as useAttendees } from './useAttendees';
-export { default as useDatetimes } from './useDatetimes';
-export { default as useEvents } from './useEvents';
-export { default as useTickets } from './useTickets';
+export * from './attendees';
+export * from './datetimes';
+export * from './tickets';
+export * from './events';
