+++ domains/blocks/src/services/apollo/queries/tickets/index.ts
@@ -1,0 +1,1 @@
+export { default as useTickets } from './useTickets';
