+++ domains/blocks/src/services/apollo/queries/tickets/queries.ts
@@ -1,0 +1,21 @@
+import gql from 'graphql-tag';
+
+export const TICKET_ATTRIBUTES: any = gql`
+	fragment blocksTicketAttributes on EspressoTicket {
+		id
+		dbId
+		cacheId
+		name
+	}
+`;
+
+export const GET_TICKETS: any = gql`
+	query GET_TICKETS($first: Int, $where: EspressoRootQueryTicketsConnectionWhereArgs) {
+		espressoTickets(first: $first, where: $where) {
+			nodes {
+				...blocksTicketAttributes
+			}
+		}
+	}
+	${TICKET_ATTRIBUTES}
+`;
